+++ src/systems/chatfilter/filters/noxd.ts
@@ -1,4 +1,7 @@
-import { Message } from "discord.js";
+import { Permissions, TextChannel, Message } from "discord.js";
+import path from "path";
+import { hasPermission } from "../../../utils/hasPermission";
+import { logHandler } from "../../../utils/logHandler";
 
 const imgUrl = "https://pics.me.me/dont-xd-350-penalty-30971828.png";
 
@@ -10,9 +13,8 @@ export default {
     async run({ message }: { message: Message }) {
         if (message.author.bot) return;
 
-        const { content, channel } = message;
+        const { guild, channel, content } = message;
         const words = content.toLowerCase().split(" ");
-
         let shouldDelete = false;
         words.forEach((word) => {
             if (table.includes(word)) {
@@ -22,6 +24,11 @@ export default {
         });
 
         if (shouldDelete) {
+            if (!hasPermission(guild.me, channel as TextChannel, [Permissions.FLAGS.MANAGE_MESSAGES])) {
+                logHandler.log("error", `${path.basename(__filename)} Dont have permission in ${channel}`);
+                return;
+            }
+
             message.delete().catch(() => { });
             const reply = await channel.send(imgUrl);
             setTimeout(() => reply.delete(), 5000);
